+++ src/Data/Suspension.hs
@@ -14,9 +14,9 @@ import Data.Data (Data)
 -- | The 'Suspension' of a type is its two-point extension:
 -- a "south pole" 'Nadir' and a "north pole" 'Zenit', if each value
 -- of the underlying type is considered as a 'Merid'ian.
-data Suspension a
+data Suspension x
   = Nadir
-  | Merid !a
+  | Merid !x
   | Zenit
   deriving
     ( Eq,
@@ -30,6 +30,20 @@ data Suspension a
       Foldable,
       Traversable
     )
+
+instance Bounded (Suspension x) where
+  minBound = Nadir
+  maxBound = Zenit
+
+instance (Ord x, Enum x, Bounded x) => Enum (Suspension x) where
+  toEnum = \case
+    0 -> Nadir
+    n | toEnum n > maxBound @x -> Zenit
+    n -> Merid (toEnum n)
+  fromEnum = \case
+    Nadir -> 0
+    Merid n -> 1 + fromEnum n
+    Zenit -> succ (fromEnum @x maxBound)
 
 instance Applicative Suspension where
   pure = Merid
@@ -91,37 +105,37 @@ instance (Num x) => Num (Suspension x) where
 --     Merid a -> merid a
 --     Zenit -> zenit
 -- @
-suspension :: b -> (a -> b) -> b -> Suspension a -> b
+suspension :: y -> (x -> y) -> y -> Suspension x -> y
 suspension nadir merid zenit = \case
   Nadir -> nadir
-  Merid a -> merid a
+  Merid x -> merid x
   Zenit -> zenit
 {-# INLINE suspension #-}
 
 -- | Unsafely extract the value from a 'Merid'.
-fromMerid :: HasCallStack => Suspension a -> a
+fromMerid :: HasCallStack => Suspension x -> x
 fromMerid = \case
   Nadir -> error "expected Merid but got Nadir!"
   Merid a -> a
   Zenit -> error "expected Merid but got Zenit!"
 
 -- | Safely extract the value from a 'Merid'.
-meridToMaybe :: Suspension a -> Maybe a
+meridToMaybe :: Suspension x -> Maybe x
 meridToMaybe = withMerid id
 {-# INLINE meridToMaybe #-}
 
 -- | Safely extract the value from a 'Merid',
 -- providing in case of failure whether the argument was 'Zenit'.
-meridToEither :: Suspension a -> Either Bool a
+meridToEither :: Suspension x -> Either Bool x
 meridToEither = suspension (Left False) Right (Left True)
 {-# INLINE meridToEither #-}
 
 -- | Apply the function to the inner value, if it exists.
-withMerid :: (Alternative m) => (a -> b) -> Suspension a -> m b
+withMerid :: (Alternative m) => (x -> y) -> Suspension x -> m y
 withMerid f = suspension empty (pure . f) empty
 {-# INLINE withMerid #-}
 
 -- | Execute the action using the inner value, if it exists.
-whenMerid :: (Alternative m) => Suspension a -> (a -> m ()) -> m ()
+whenMerid :: (Alternative m) => Suspension x -> (x -> m ()) -> m ()
 whenMerid s f = suspension pass f pass s
 {-# INLINE whenMerid #-}